requirements.txt
alembic==1.8.1
autopep8==1.5.7
Babel==2.10.3
bcrypt==3.2.2
blinker==1.5
boto==2.49.0
boto3==1.24.83
cffi==1.15.1
click==8.1.3
colorama==0.4.5
cryptography==38.0.1
dnspython==2.2.1
email-validator==1.3.0
Flask==2.2.2
Flask-Admin==1.6.0
flask-admin-s3-upload==0.1.4
Flask-BabelEx==0.9.4
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
Flask-Mail==0.9.1
flask-marshmallow==0.14.0
Flask-Migrate==3.1.0
Flask-Principal==0.4.0
Flask-Security==3.0.0
Flask-Security-Too==5.0.2
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.2
gunicorn==20.1.0
idna==3.4
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.2
MarkupSafe==2.1.1
marshmallow==3.17.0
marshmallow-sqlalchemy==0.28.1
packaging==21.3
passlib==1.7.4
Pillow==9.2.0
psycopg2-binary==2.9.3
pycodestyle==2.7.0
pycparser==2.21
PyJWT==2.5.0
pyparsing==3.0.9
python-dotenv==0.20.0
python3-pip-autoremove==0.9.1
pytz==2022.2.1
six==1.16.0
speaklater==1.3
SQLAlchemy==1.4.40
url-for-s3==0.1.2
Werkzeug==2.2.2
wincertstore==0.2
WTForms==3.0.1
# psycopg2
